+++ CHANGELOG.md
@@ -8,8 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Drop Python 3.6 support ([#319](https://github.com/warpnet/salt-lint/pull/319)).
 
 ### Added
-- Add Python 3.13 and 3.14 support.
-- Add Python 3.12 support ([#315](https://github.com/warpnet/salt-lint/pull/315)).
+- Add `--codeclimate` output format for GitLab Code Quality reports ([#330](https://github.com/warpnet/salt-lint/issues/330)).
+- Add Python 3.12, 3.13 and 3.14 support ([#315](https://github.com/warpnet/salt-lint/pull/315), [#331](https://github.com/warpnet/salt-lint/pull/331)).
 - Lookup configuration file in parent directory ([#305](https://github.com/warpnet/salt-lint/pull/305)).
 
 ### Fixed